todoist: hold todo until request queue is empty
This patch introduces a hold-state which should
be active when todoist request queue is not empty.
User can accidently close the app when there are
requests yet to be dispatched, this can cause lose
of request. With this patch todo continues to run
in background until all the post requests are
dispatched.

https://bugzilla.gnome.org/show_bug.cgi?id=772278