project: Fix documentation generation
Documentation generation with GtkDoc is failing due to missing
directories.

This is happening because GtkDoc tries to search for directories
that are present in the source code tree, but are not in the build
tree: 'src/engine', 'src/interfaces' and 'src/notification'.

This patch modifies the directories that GtkDoc scans in search for
source code files.

https://bugzilla.gnome.org/show_bug.cgi?id=788295