task-list-view: Done button has maximum width
This patch adds a maximum width to the Done button in
the task list view.

A subclass of the GtkButton is created. This class is
named GtdDoneButton. In this class the get_preferred_width()
method is overrided by gtd_row_get_preferred_width_with_max()
from gtd-commons.

The type of the Done button is changed in list-view.ui from
GtkButton to GtdDoneButton and also the halign(center) property
is added.