object: Change semantics of ready to loading
Instead of explicit set, use push/pop pairs to make it
more robust.

The super nice side effect is that this commit also fixes
the old loading notification, that was actually never removed
but just stopped working due to the backend reworks.

This also allowed us to remove and old EDS hack.