night-light: Introduce Night Light plugin
This is an enabled-by-default plugin that monitors Night Light
and switches to the dark theme variant when Night Light is on.

It does some effort to return to the correct theme variant after,
but not too much.

It can be disabled by editing the list of enabled plugins with
D-Conf editor.