task-eds: Add reversible editing
This will be useful in the future. GtdTaskEds now stores
the original component, and redirects all editing into
a new one.

When an update is successful, this new component becomes
the actual one, and the process repeats (another component
is created, all editing is redirected, etc).

On error, however, the original component is restored and
we can at least show the real contents of the task, instead
of pretending everything went fine.