project: Fix list model constructors' ownership
Before GTK 3.99, the various types of list models that
GTK offer had transfer-none constructors. Meaning, they
usually had to be used in conjunction to g_autoptr or
g_object_unref.

Since commit 706d464ae6d5 [1], however, the constructors
are transfer-full, and they don't take a ref on the passed
objects. The intention seems to be to allow the list model
APIs to be used as:

    gtk_filter_list_model_new (model, gtk_filter_new (…));

Update GNOME To Do to not unref the GtkFilters and
GtkSorters when passing them to the models.

[1] https://gitlab.gnome.org/GNOME/gtk/-/commit/706d464ae6d5d3f120a2fb22932f49e07f034545