Introduce an animation framework
This is pretty much a direct port of Clutter's animation framework,
except it uses GdkFrameClock internally. The idea is to make all
widgetry of GNOME To Do use it.

It supports both implicit and explicit easing; has property- and
keyframe-based transitions; and supports setting custom markers
at timelines.

Also add a new interactive test to showcase it.